chore: restructure modules
